s2e.so: fixed process memory map address ranges
The calculation of the "limit" address for an entry in the process memory map
was incorrect. The previous method only looked at executable entries in the
process memory map. However, it is common for non-executable entries in the
same virtual address space to follow the executable entry.

The process memory map parsing logic was updated to handle these cases. The
parsing code now uses inodes to determine new entries in the process memory
map. While the inode remains unchanged, the "limit" address is checked to
ensure that the maximum address is saved into the final process memory map.

This assumes that entries of the same inode share a contiguous virtual address
space.